Adding new stuff for gameboardextended creation
