+++ .replit
@@ -1,2 +1,2 @@
 language = "go"
-run = "go run main.go logic.go"
+run = "go run ."
