+++ main.go
@@ -7,70 +7,6 @@ import (
 	"os"
 )
 
-type ContentType int
-
-const (
-	Food ContentType = iota
-	Head
-	Body
-	Empty
-)
-
-type CoordExtended struct {
-	score   int
-	content ContentType
-}
-
-type GameBoardExtended [][]CoordExtended
-
-// --
-
-type GameState struct {
-	Game  Game        `json:"game"`
-	Turn  int         `json:"turn"`
-	Board Board       `json:"board"`
-	You   Battlesnake `json:"you"`
-}
-
-type Game struct {
-	ID      string  `json:"id"`
-	Ruleset Ruleset `json:"ruleset"`
-	Timeout int32   `json:"timeout"`
-}
-
-type Ruleset struct {
-	Name    string `json:"name"`
-	Version string `json:"version"`
-}
-
-type Board struct {
-	Height int           `json:"height"`
-	Width  int           `json:"width"`
-	Food   []Coord       `json:"food"`
-	Snakes []Battlesnake `json:"snakes"`
-
-	// Used in non-standard game modes
-	Hazards []Coord `json:"hazards"`
-}
-
-type Battlesnake struct {
-	ID      string  `json:"id"`
-	Name    string  `json:"name"`
-	Health  int32   `json:"health"`
-	Body    []Coord `json:"body"`
-	Head    Coord   `json:"head"`
-	Length  int32   `json:"length"`
-	Latency string  `json:"latency"`
-
-	// Used in non-standard game modes
-	Shout string `json:"shout"`
-	Squad string `json:"squad"`
-}
-
-type Coord struct {
-	X int `json:"x"`
-	Y int `json:"y"`
-}
 
 // Response Structs
